+++ test/integration/triggerTracesTest.js
@@ -116,7 +116,5 @@ describe.skip('TriggerTrace', function() {
         should(response).have.property('prev').which.is.null();
       });
     });
-
   });
-
 });